Bluetooth: Store the source address type of LE connections
When establishing LE connections, it is possible to use a public
address (if available) or a random address. The type of address
is only known when creating connections, so make sure it is
stored in hci_conn structure.

Change-Id: I31d0a91311b34d16b79b97d873e44dd6a9c23c35
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
Signed-off-by: Johan Hedberg <johan.hedberg@intel.com>
Git-commit: e7c4096e16f0e362c6cf902baab0de37ebfc1266
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Signed-off-by: Kamal Negi <kamaln@codeaurora.org>
Signed-off-by: Laster K. (lazerl0rd) <officiallazerl0rd@gmail.com>